Split OpenTelemetry result handler

Copilot-Session: 78c72398-8e6e-49c1-bf8d-9be36e921d9e
