+++ src/Platform/Microsoft.Testing.Platform/Telemetry/OpenTelemetryResultHandler.cs
@@ -2,21 +2,11 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Testing.Platform.Extensions.Messages;
-using Microsoft.Testing.Platform.Helpers;
 
 namespace Microsoft.Testing.Platform.Telemetry;
 
-internal sealed class OpenTelemetryResultHandler : IDisposable
+internal sealed partial class OpenTelemetryResultHandler : IDisposable
 {
-    /// <summary>
-    /// What Roslyn's <c>INamespaceSymbol.ToDisplayString()</c> returns for the global namespace.
-    /// </summary>
-    private const string RoslynGlobalNamespaceDisplayString = "<global namespace>";
-    private const string ExceptionEventName = "exception";
-    private const string ExceptionTypeTag = "exception.type";
-    private const string ExceptionMessageTag = "exception.message";
-    private const string ExceptionStackTraceTag = "exception.stacktrace";
-
     private readonly IPlatformOpenTelemetryService _otelService;
     private readonly PlatformOpenTelemetryOptions _options;
 
@@ -100,105 +90,6 @@ internal sealed class OpenTelemetryResultHandler : IDisposable
         }
     }
 
-    internal void NotifyDiscovered()
-        => _totalDiscoveredTests?.Add(1);
-
-    internal void NotifyPassed(TestNode testNode, TestNodeStateProperty stateProperty)
-    {
-        _totalPassedTests?.Add(1);
-        HandleTestResult(testNode, stateProperty);
-    }
-
-    internal void NotifyFailed(TestNode testNode, TestNodeStateProperty stateProperty)
-    {
-        _totalFailedTests?.Add(1);
-        HandleTestResult(testNode, stateProperty);
-    }
-
-    internal void NotifySkipped(TestNode testNode, TestNodeStateProperty stateProperty)
-    {
-        _totalSkippedTests?.Add(1);
-        HandleTestResult(testNode, stateProperty);
-    }
-
-    internal void NotifyInProgress(TestNode testNode, TestNodeUid? parentUid)
-    {
-        _totalStartedTests?.Add(1);
-        IPlatformActivity? activity = _otelService.StartActivity(
-            GetActivityName(testNode),
-            parentId: _otelService.TestFrameworkActivity?.Id,
-            tags: GetTestInitialInfo(testNode, parentUid));
-
-        lock (_syncRoot)
-        {
-            if (_disposed)
-            {
-                // A result arrived after shutdown started; close its span rather than tracking it.
-                activity?.Dispose();
-                return;
-            }
-
-            _activeTestCases.Add(1);
-            if (!_testActivities.TryGetValue(testNode.Uid, out Queue<IPlatformActivity?>? activities))
-            {
-                activities = new Queue<IPlatformActivity?>();
-                _testActivities.Add(testNode.Uid, activities);
-            }
-
-            activities.Enqueue(activity);
-        }
-    }
-
-    internal void NotifyExecutionCompleted(TestNode testNode)
-    {
-        _totalCompletedTests?.Add(1);
-        if (!TryDequeueInFlight(testNode, out IPlatformActivity? activity))
-        {
-            return;
-        }
-
-        activity?.Dispose();
-    }
-
-    internal void NotifyUnknown()
-        => _totalUnknownTests?.Add(1);
-
-    /// <summary>
-    /// Records the run-level metrics. Called once, when the run verdict is known.
-    /// </summary>
-    /// <param name="totalRanTests">Number of tests that ran.</param>
-    /// <param name="failedTests">Number of tests that failed.</param>
-    /// <param name="skippedTests">Number of tests that were skipped.</param>
-    /// <param name="exitCode">The process exit code the run resolved to.</param>
-    /// <param name="runActivity">The root span of the run, if any. The counts go here rather than on the histogram
-    /// because they are unbounded values: as metric dimensions they would create a new time series per distinct
-    /// count, while on a span they are free.</param>
-    internal void NotifyRunCompleted(int totalRanTests, int failedTests, int skippedTests, int exitCode, IPlatformActivity? runActivity = null)
-    {
-        // Dispose can legitimately run more than once; recording a second data point would double count the run.
-        lock (_syncRoot)
-        {
-            if (_runCompletedReported)
-            {
-                return;
-            }
-
-            _runCompletedReported = true;
-        }
-
-        _runStopwatch.Stop();
-        _testRunDuration.Record(
-            _runStopwatch.Elapsed.TotalSeconds,
-            [
-                new(TestingPlatformSemanticConventions.Attributes.TestRunResultStatus, GetRunResultStatus(failedTests, exitCode)),
-                new(TestingPlatformSemanticConventions.Attributes.TestRunExitCode, exitCode),
-            ]);
-
-        runActivity?.SetTag(TestingPlatformSemanticConventions.Attributes.TestRunTotalCount, totalRanTests);
-        runActivity?.SetTag(TestingPlatformSemanticConventions.Attributes.TestRunFailedCount, failedTests);
-        runActivity?.SetTag(TestingPlatformSemanticConventions.Attributes.TestRunSkippedCount, skippedTests);
-    }
-
     public void Dispose()
     {
         List<IPlatformActivity?> orphaned = [];
@@ -229,379 +120,6 @@ internal sealed class OpenTelemetryResultHandler : IDisposable
         foreach (IPlatformActivity? activity in orphaned)
         {
             activity?.Dispose();
-        }
-    }
-
-    /// <summary>
-    /// Removes the oldest in-flight entry for the node and keeps <c>test.case.active</c> balanced.
-    /// </summary>
-    private bool TryDequeueInFlight(TestNode testNode, out IPlatformActivity? activity)
-    {
-        lock (_syncRoot)
-        {
-            activity = null;
-            if (!_testActivities.TryGetValue(testNode.Uid, out Queue<IPlatformActivity?>? activities) || activities.Count == 0)
-            {
-                return false;
-            }
-
-            activity = activities.Dequeue();
-            if (activities.Count == 0)
-            {
-                _testActivities.Remove(testNode.Uid);
-            }
-
-            _activeTestCases.Add(-1);
-            return true;
-        }
-    }
-
-    /// <summary>
-    /// The OpenTelemetry conventions ask for the span name to be the test case name rather than an opaque id,
-    /// because it is what shows up in trace waterfalls and what backends group on.
-    /// </summary>
-    private static string GetActivityName(TestNode testNode)
-        => RoslynString.IsNullOrWhiteSpace(testNode.DisplayName) ? testNode.Uid.Value : testNode.DisplayName;
-
-    private static string? GetSuiteName(TestNode testNode)
-        => testNode.Properties.SingleOrDefault<TestMethodIdentifierProperty>()?.TypeName;
-
-    /// <summary>
-    /// Builds the value for <c>code.function.name</c>, which the convention defines as the fully qualified name.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// A type in the global namespace contributes no namespace segment, so the name must not gain a leading dot.
-    /// Two spellings have to be recognised, because the property is a plain string filled in by whichever test
-    /// framework produced the node: an empty namespace (what MSTest produces, since it derives the namespace by
-    /// splitting the managed type name) and Roslyn's <c>&lt;global namespace&gt;</c>, which is what
-    /// <c>INamespaceSymbol.ToDisplayString()</c> returns when the caller does not first check
-    /// <c>IsGlobalNamespace</c>.
-    /// </para>
-    /// <para>
-    /// Dropping the segment - rather than emitting Roslyn's <c>global::</c> prefix - keeps the attribute
-    /// language-agnostic, matches the convention's own examples, and is consistent with how the rest of this
-    /// repository treats the global namespace (see <c>TestClassModelBuilder</c>, <c>ReflectionMetadataGenerator</c>
-    /// and <c>DiscoveredTestsJsonSerializer</c>, which all omit it).
-    /// </para>
-    /// </remarks>
-    private static string GetFullyQualifiedName(TestMethodIdentifierProperty identifierProperty)
-        => IsGlobalNamespace(identifierProperty.Namespace)
-            ? $"{identifierProperty.TypeName}.{identifierProperty.MethodName}"
-            : $"{identifierProperty.Namespace}.{identifierProperty.TypeName}.{identifierProperty.MethodName}";
-
-    private static bool IsGlobalNamespace(string? @namespace)
-        => RoslynString.IsNullOrEmpty(@namespace)
-            || @namespace == RoslynGlobalNamespaceDisplayString;
-
-    private static string GetRunResultStatus(int failedTests, int exitCode)
-        => failedTests > 0 || exitCode != (int)ExitCode.Success
-            ? TestingPlatformSemanticConventions.TestResultStatus.Fail
-            : TestingPlatformSemanticConventions.TestResultStatus.Pass;
-
-    private IEnumerable<KeyValuePair<string, object?>> GetTestInitialInfo(TestNode testNode, TestNodeUid? parentUid)
-    {
-        yield return new(TestingPlatformSemanticConventions.Attributes.TestCaseName, testNode.DisplayName);
-        yield return new(TestingPlatformSemanticConventions.Attributes.TestCaseId, testNode.Uid.Value);
-        if (parentUid is not null)
-        {
-            yield return new(TestingPlatformSemanticConventions.Attributes.TestCaseParentId, parentUid.Value);
-        }
-
-        if (_options.EmitLegacyAttributes)
-        {
-            yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestName, testNode.DisplayName);
-            yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestId, testNode.Uid.Value);
-            if (parentUid is not null)
-            {
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestParentId, parentUid.Value);
-            }
-        }
-
-        // Single pass over the property bag for the two singleton properties below: two separate
-        // SingleOrDefault<T>() calls walked the whole bag once each.
-        TestMethodIdentifierProperty? identifierProperty = null;
-        TestFileLocationProperty? testLocationProperty = null;
-        PropertyBag.PropertyBagEnumerator enumerator = testNode.Properties.GetStructEnumerator();
-        while (enumerator.MoveNext())
-        {
-            switch (enumerator.Current)
-            {
-                case TestMethodIdentifierProperty identifier:
-                    if (identifierProperty is not null)
-                    {
-                        throw new InvalidOperationException($"Found multiple properties of type '{typeof(TestMethodIdentifierProperty)}'.");
-                    }
-
-                    identifierProperty = identifier;
-                    break;
-                case TestFileLocationProperty location:
-                    if (testLocationProperty is not null)
-                    {
-                        throw new InvalidOperationException($"Found multiple properties of type '{typeof(TestFileLocationProperty)}'.");
-                    }
-
-                    testLocationProperty = location;
-                    break;
-            }
-        }
-
-        if (identifierProperty is not null)
-        {
-            // code.function.name is defined as the *fully qualified* name; there is no separate namespace
-            // attribute (code.namespace is deprecated upstream).
-            yield return new(TestingPlatformSemanticConventions.Attributes.CodeFunctionName, GetFullyQualifiedName(identifierProperty));
-            yield return new(TestingPlatformSemanticConventions.Attributes.TestSuiteName, identifierProperty.TypeName);
-            yield return new(TestingPlatformSemanticConventions.Attributes.TestAssemblyName, identifierProperty.AssemblyFullName);
-
-            if (_options.EmitLegacyAttributes)
-            {
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestMethod, identifierProperty.MethodName);
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestClass, identifierProperty.TypeName);
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestNamespace, identifierProperty.Namespace);
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestAssembly, identifierProperty.AssemblyFullName);
-            }
-        }
-
-        if (testLocationProperty is not null)
-        {
-            yield return new(TestingPlatformSemanticConventions.Attributes.CodeFilePath, testLocationProperty.FilePath);
-            yield return new(TestingPlatformSemanticConventions.Attributes.CodeLineNumber, testLocationProperty.LineSpan.Start.Line);
-
-            if (_options.EmitLegacyAttributes)
-            {
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestFilePath, testLocationProperty.FilePath);
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestLineStart, testLocationProperty.LineSpan.Start.Line);
-                yield return new(TestingPlatformSemanticConventions.Attributes.LegacyTestLineEnd, testLocationProperty.LineSpan.End.Line);
-            }
-        }
-
-        // The metadata is yielded after the blocks above, so it needs its own pass. OfType<TestMetadataProperty>()
-        // would materialize a TProperty[] just to enumerate it once; the struct enumerator allocates nothing.
-        PropertyBag.PropertyBagEnumerator metadataEnumerator = testNode.Properties.GetStructEnumerator();
-        while (metadataEnumerator.MoveNext())
-        {
-            if (metadataEnumerator.Current is not TestMetadataProperty metadata)
-            {
-                continue;
-            }
-
-            yield return new KeyValuePair<string, object?>($"{TestingPlatformSemanticConventions.Attributes.TestMetadataPrefix}{metadata.Key}", metadata.Value);
-            if (_options.EmitLegacyAttributes)
-            {
-                yield return new KeyValuePair<string, object?>($"{TestingPlatformSemanticConventions.Attributes.LegacyTestMetadataPrefix}{metadata.Key}", metadata.Value);
-            }
-        }
-    }
-
-    private void HandleTestResult(TestNode testNode, TestNodeStateProperty stateProperty)
-    {
-        _totalCompletedTests?.Add(1);
-
-        (string result, Exception? exception, TimeSpan? timeoutTime) = stateProperty switch
-        {
-            PassedTestNodeStateProperty => (TestingPlatformSemanticConventions.TestResultStatus.Pass, null, null),
-            FailedTestNodeStateProperty failed => (TestingPlatformSemanticConventions.TestResultStatus.Fail, failed.Exception, null),
-            ErrorTestNodeStateProperty error => (TestingPlatformSemanticConventions.TestResultStatus.Error, error.Exception, null),
-            TimeoutTestNodeStateProperty timeout => (TestingPlatformSemanticConventions.TestResultStatus.Timeout, timeout.Exception, timeout.Timeout),
-#pragma warning disable CS0618, MTP0001 // Type or member is obsolete
-            CancelledTestNodeStateProperty cancelled => (TestingPlatformSemanticConventions.TestResultStatus.Cancelled, cancelled.Exception, null),
-#pragma warning restore CS0618, MTP0001 // Type or member is obsolete
-            SkippedTestNodeStateProperty => (TestingPlatformSemanticConventions.TestResultStatus.Skipped, null, null),
-            _ => (TestingPlatformSemanticConventions.TestResultStatus.Unknown, null, null),
-        };
-
-        KeyValuePair<string, object?>[] measurementTags =
-        [
-            new(TestingPlatformSemanticConventions.Attributes.TestCaseResultStatus, result),
-            new(TestingPlatformSemanticConventions.Attributes.TestSuiteName, GetSuiteName(testNode)),
-        ];
-
-        _testCaseResultCount.Add(1, measurementTags);
-
-        if (!TryDequeueInFlight(testNode, out IPlatformActivity? activity) || activity is null)
-        {
-            // Either the framework never reported the test as in-progress, or nothing is listening so no span was
-            // created. Either way we still want the duration recorded, otherwise a framework that only publishes
-            // final results produces no latency data.
-            SetResultDetails(testNode, measurementTags, activity: null);
-            return;
-        }
-
-        string? truncatedExplanation = _options.Truncate(stateProperty.Explanation);
-        activity.SetTag(TestingPlatformSemanticConventions.Attributes.TestCaseResultStatus, result);
-        activity.SetTag(TestingPlatformSemanticConventions.Attributes.TestCaseResultExplanation, truncatedExplanation);
-
-        if (_options.EmitLegacyAttributes)
-        {
-            // The legacy attribute keeps its original "passed"/"failed" spellings; the semantic-convention
-            // attribute uses the upstream "pass"/"fail" enum.
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResult, TestingPlatformSemanticConventions.TestResultStatus.ToLegacy(result));
-
-            // Truncated as well: emitting the legacy twin untruncated would defeat the size limit, since legacy
-            // attributes are on by default.
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResultExplanation, truncatedExplanation);
-        }
-
-        if (exception is not null)
-        {
-            // The OpenTelemetry convention is an "exception" event carrying the type/message/stack trace, plus
-            // error.type and a status of Error on the span so it shows up as failed in every backend.
-            // error.message is deliberately not set: it is deprecated upstream and NOT RECOMMENDED on spans
-            // because of its unbounded cardinality - the message is on the event instead.
-            string? exceptionTypeName = exception.GetType().FullName;
-            string? truncatedMessage = _options.Truncate(exception.Message);
-            string? truncatedStackTrace = _options.Truncate(exception.StackTrace);
-            string? truncatedExceptionStackTrace = _options.Truncate(exception.ToString());
-
-            activity.AddEvent(
-                ExceptionEventName,
-                [
-                    new(ExceptionTypeTag, exceptionTypeName),
-                    new(ExceptionMessageTag, truncatedMessage),
-                    new(ExceptionStackTraceTag, truncatedExceptionStackTrace),
-                ]);
-            activity.SetStatus(PlatformActivityStatusCode.Error, truncatedMessage);
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.ErrorType, exceptionTypeName);
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.CodeStacktrace, truncatedStackTrace);
-
-            if (_options.EmitLegacyAttributes)
-            {
-                activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResultExceptionType, exceptionTypeName);
-                activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResultExceptionMessage, truncatedMessage);
-                activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResultExceptionStackTrace, truncatedStackTrace);
-            }
-        }
-        else
-        {
-            activity.SetStatus(
-                result switch
-                {
-                    TestingPlatformSemanticConventions.TestResultStatus.Pass => PlatformActivityStatusCode.Ok,
-                    TestingPlatformSemanticConventions.TestResultStatus.Skipped or TestingPlatformSemanticConventions.TestResultStatus.Unknown => PlatformActivityStatusCode.Unset,
-                    _ => PlatformActivityStatusCode.Error,
-                },
-                truncatedExplanation);
-        }
-
-        if (timeoutTime is not null)
-        {
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.TestCaseTimeoutMilliseconds, timeoutTime.Value.TotalMilliseconds);
-            if (_options.EmitLegacyAttributes)
-            {
-                activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestResultTimeout, timeoutTime.Value.TotalMilliseconds);
-            }
-        }
-
-        try
-        {
-            SetResultDetails(testNode, measurementTags, activity);
-        }
-        finally
-        {
-            // The span was already dequeued, so it must be closed even if collecting the details throws on a
-            // malformed property bag; otherwise it would stay open until the handler is disposed.
-            activity.Dispose();
-        }
-    }
-
-    /// <summary>
-    /// Collects the timing, output and artifact details of a completed test in a single pass over the property bag.
-    /// </summary>
-    private void SetResultDetails(TestNode testNode, KeyValuePair<string, object?>[] measurementTags, IPlatformActivity? activity)
-    {
-        // Single pass over the property bag: replaces five separate walks
-        // (SingleOrDefault<TimingProperty>, OfType<TestMetadataProperty>, SingleOrDefault<StandardOutputProperty>,
-        //  SingleOrDefault<StandardErrorProperty>, OfType<FileArtifactProperty>).
-        // Eliminates two TProperty[] heap allocations from OfType<T>() and reduces linked-list traversal from 5 to 1.
-        TimingProperty? timingProperty = null;
-        StandardOutputProperty? standardOutputProperty = null;
-        StandardErrorProperty? standardErrorProperty = null;
-        int artifactIndex = 0;
-        PropertyBag.PropertyBagEnumerator enumerator = testNode.Properties.GetStructEnumerator();
-        while (enumerator.MoveNext())
-        {
-            switch (enumerator.Current)
-            {
-                case TimingProperty tp:
-                    if (timingProperty is not null)
-                    {
-                        throw new InvalidOperationException($"Found multiple properties of type '{typeof(TimingProperty)}'.");
-                    }
-
-                    timingProperty = tp;
-                    break;
-                case TestMetadataProperty metadataProperty:
-                    activity?.SetTag($"{TestingPlatformSemanticConventions.Attributes.TestMetadataPrefix}{metadataProperty.Key}", metadataProperty.Value);
-                    if (_options.EmitLegacyAttributes)
-                    {
-                        activity?.SetTag($"{TestingPlatformSemanticConventions.Attributes.LegacyTestMetadataPrefix}{metadataProperty.Key}", metadataProperty.Value);
-                    }
-
-                    break;
-                case StandardOutputProperty outputProperty:
-                    if (standardOutputProperty is not null)
-                    {
-                        throw new InvalidOperationException($"Found multiple properties of type '{typeof(StandardOutputProperty)}'.");
-                    }
-
-                    standardOutputProperty = outputProperty;
-                    break;
-                case StandardErrorProperty errorProperty:
-                    if (standardErrorProperty is not null)
-                    {
-                        throw new InvalidOperationException($"Found multiple properties of type '{typeof(StandardErrorProperty)}'.");
-                    }
-
-                    standardErrorProperty = errorProperty;
-                    break;
-                case FileArtifactProperty fileArtifactProperty:
-                    activity?.SetTag($"test.artifact.file[{artifactIndex}].path", fileArtifactProperty.FileInfo.FullName);
-                    artifactIndex++;
-                    break;
-            }
-        }
-
-        if (timingProperty is not null)
-        {
-            double totalMilliseconds = timingProperty.GlobalTiming.Duration.TotalMilliseconds;
-            _testCaseDuration.Record(timingProperty.GlobalTiming.Duration.TotalSeconds, measurementTags);
-            _totalDuration?.Record(totalMilliseconds);
-            activity?.SetTag(TestingPlatformSemanticConventions.Attributes.TestCaseDurationMilliseconds, totalMilliseconds);
-            if (_options.EmitLegacyAttributes)
-            {
-                activity?.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestDuration, totalMilliseconds);
-            }
-
-            foreach (StepTimingInfo step in timingProperty.StepTimings)
-            {
-                activity?.SetTag($"{TestingPlatformSemanticConventions.Attributes.TestStepPrefix}{step.Id}.duration", step.Timing.Duration.TotalMilliseconds);
-                activity?.SetTag($"{TestingPlatformSemanticConventions.Attributes.TestStepPrefix}{step.Id}.description", step.Description);
-                if (_options.EmitLegacyAttributes)
-                {
-                    activity?.SetTag($"test.step{step.Id}.duration.ms", step.Timing.Duration.TotalMilliseconds);
-                    activity?.SetTag($"test.step{step.Id}.description", step.Description);
-                }
-            }
-        }
-
-        if (activity is null || !_options.CaptureTestOutput)
-        {
-            return;
-        }
-
-        // Truncated for both the semantic-convention and the legacy names: test output routinely runs to megabytes
-        // and can contain secrets.
-        string standardOutput = _options.Truncate(standardOutputProperty?.StandardOutput) ?? string.Empty;
-        string standardError = _options.Truncate(standardErrorProperty?.StandardError) ?? string.Empty;
-
-        activity.SetTag(TestingPlatformSemanticConventions.Attributes.TestOutputStdout, standardOutput);
-        activity.SetTag(TestingPlatformSemanticConventions.Attributes.TestOutputStderr, standardError);
-
-        if (_options.EmitLegacyAttributes)
-        {
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestStdout, standardOutput);
-            activity.SetTag(TestingPlatformSemanticConventions.Attributes.LegacyTestStderr, standardError);
         }
     }
 }
